Enables windows symlinking for collect_deploy_content
Tested with base_test and verified that symlink
was created.

out/nxswitch_debug/deploy/base_test/icu (link)
-> maps to ->
out/nxswitch_debug/deploy/content/deploy/base_test/icu

+++ starboard/build/collect_deploy_content.py
@@ -22,9 +22,18 @@ import os
 import shutil
 import sys
 
+import _env # pylint: disable=unused-import
+
 # The name of an environment variable that when set to |'1'|, signals to us that
 # we should log all output directories that we have populated.
 _SHOULD_LOG_ENV_KEY = 'STARBOARD_GYP_SHOULD_LOG_COPIES'
+
+
+_USE_WINDOWS_SYMLINK = sys.platform in ['win32', 'cygwin'] and \
+                       not hasattr(os, 'symlink')
+
+if _USE_WINDOWS_SYMLINK:
+  import starboard.build.win_symlink as win_symlink
 
 
 def EscapePath(path):
@@ -58,7 +67,10 @@ def main(argv):
     logging.info('+ %s', subdir)
 
   if os.path.exists(options.output_dir):
-    shutil.rmtree(options.output_dir)
+    if _USE_WINDOWS_SYMLINK:
+      win_symlink.Rmtree(options.output_dir)
+    else:
+      shutil.rmtree(options.output_dir)
 
   for subdir in options.subdirs:
     src_path = os.path.abspath(
@@ -71,14 +83,18 @@ def main(argv):
 
     logging.info('%s => %s', dst_path, rel_path)
 
-    # TODO: Add an alternate implementation for win32.
     if not os.path.exists(dst_dir):
       os.makedirs(dst_dir)
-    os.symlink(rel_path, dst_path)
+
+    if _USE_WINDOWS_SYMLINK:
+      win_symlink.CreateReparsePoint(src_path, dst_path)
+    else:
+      os.symlink(rel_path, dst_path)
 
   if options.stamp_file:
     with open(options.stamp_file, 'w') as stamp_file:
       stamp_file.write('\n'.join(options.subdirs))
+
 
 if __name__ == '__main__':
   sys.exit(main(sys.argv))
